Now the Offer help page is showing help offers created by user Co-authored-by: Damaso Junio <junio.pereirab@gmail.com> Co-authored-by: Pedro Leamon <phlimas@outlook.com>
